+++ crates/biome_js_analyze/src/lib.rs
@@ -94,7 +94,11 @@ where
                 } else {
                     let category = key.name();
                     if let Some(rule) = category.strip_prefix("lint/") {
-                        result.push(Ok(SuppressionKind::Rule(rule)));
+                        if let Some(value) = value {
+                            result.push(Ok(SuppressionKind::RuleWithValue(rule, value)));
+                        } else {
+                            result.push(Ok(SuppressionKind::Rule(rule)));
+                        }
                     }
                 }
             }